Debug

+++ models/hrm/hrm_act_v1.py
@@ -183,7 +183,7 @@ class HierarchicalReasoningModel_ACTV1_Inner(nn.Module):
         z_H = self.H_level(z_H, z_L, **seq_info)
 
         new_carry = HierarchicalReasoningModel_ACTV1InnerCarry(z_H=z_H.detach(), z_L=z_L.detach())
-        output = self.lm_head(z_H)[:, self.puzzle_emb_len:]
+        output = self.lm_head(z_H.float())[:, self.puzzle_emb_len:].to(z_H.dtype)
 
         q_logits = self.q_head(z_H[:, 0]).to(torch.float32)
